deepnet/experiments/kerasnet/requirements.txt
tensorflow_gpu~=2.10.0
keras~=2.10.0
